creating clouds

+++ core/src/helpers/GameInfo.java
@@ -5,5 +5,5 @@ public class GameInfo {
     public static final int WIDTH = 480;
     public static final int HEIGHT = 800;
     //pixels per meter: 1 meter = 100 pixels
-    public static final int PPM = 100;
+    public static final float PPM = 100;
 }
